+++ src/main/java/net/glowstone/generator/decorators/overworld/CactusDecorator.java
@@ -3,11 +3,10 @@ package net.glowstone.generator.decorators.overworld;
 import java.util.Random;
 
 import net.glowstone.generator.decorators.BlockDecorator;
+import net.glowstone.generator.objects.Cactus;
 
 import org.bukkit.Chunk;
-import org.bukkit.Material;
 import org.bukkit.World;
-import org.bukkit.block.Block;
 
 public class CactusDecorator extends BlockDecorator {
 
@@ -21,23 +20,7 @@ public class CactusDecorator extends BlockDecorator {
             int x = sourceX + random.nextInt(8) - random.nextInt(8);
             int z = sourceZ + random.nextInt(8) - random.nextInt(8);
             int y = sourceY + random.nextInt(4) - random.nextInt(4);
-
-            if (world.getBlockAt(x, y, z).getType() == Material.AIR) {
-
-                int height = random.nextInt(random.nextInt(3) + 1) + 1;
-                for (int n = y; n < y + height; n++) {
-                    final Block block = world.getBlockAt(x, n, z);
-                    if ((world.getBlockAt(x, n - 1, z).getType() == Material.SAND ||
-                            world.getBlockAt(x, n - 1, z).getType() == Material.CACTUS) &&
-                            world.getBlockAt(x, n + 1, z).getType() == Material.AIR &&
-                            !world.getBlockAt(x - 1, n, z).getType().isSolid() &&
-                            !world.getBlockAt(x + 1, n, z).getType().isSolid() &&
-                            !world.getBlockAt(x, n, z - 1).getType().isSolid() &&
-                            !world.getBlockAt(x, n, z + 1).getType().isSolid())
-                    block.setType(Material.CACTUS);
-                    block.setData((byte) 0);
-                }
-            }
+            new Cactus().generate(world, random, x, y, z);
         }
     }
 }
