+++ src/main/java/net/glowstone/generator/populators/OrePopulator.java
@@ -1,8 +1,12 @@
 package net.glowstone.generator.populators;
 
+import java.util.LinkedHashMap;
+import java.util.Map;
 import java.util.Random;
+import java.util.Map.Entry;
 
-import net.glowstone.generator.OreVeinGenerator;
+import net.glowstone.generator.objects.OreType;
+import net.glowstone.generator.objects.OreVein;
 
 import org.bukkit.Chunk;
 import org.bukkit.Material;
@@ -14,24 +18,40 @@ import org.bukkit.generator.BlockPopulator;
  */
 public class OrePopulator extends BlockPopulator {
 
-    private final OreVeinGenerator generator = new OreVeinGenerator();
+    private final Map<OreType, Integer> ores = new LinkedHashMap<>();
 
     public OrePopulator() {
-        generator.addOre(10, Material.DIRT, 0, 256, 32);
-        generator.addOre(8, Material.GRAVEL, 0, 256, 32);
-        generator.addOre(10, Material.STONE, 1, 0, 80, 32);
-        generator.addOre(10, Material.STONE, 3, 0, 80, 32);
-        generator.addOre(10, Material.STONE, 5, 0, 80, 32);
-        generator.addOre(20, Material.COAL_ORE, 0, 128, 16);
-        generator.addOre(20, Material.IRON_ORE, 0, 64, 8);
-        generator.addOre(2, Material.GOLD_ORE, 0, 32, 8);
-        generator.addOre(8, Material.REDSTONE_ORE, 0, 16, 7);
-        generator.addOre(1, Material.DIAMOND_ORE, 0, 16, 7);
-        generator.addOre(1, Material.LAPIS_ORE, 16, 16, 6);
+        ores.put(new OreType(Material.DIRT, 0, 0, 256, 32), 10);
+        ores.put(new OreType(Material.GRAVEL, 0, 0, 256, 32), 8);
+        ores.put(new OreType(Material.STONE, 1, 0, 80, 32), 10);
+        ores.put(new OreType(Material.STONE, 3, 0, 80, 32), 10);
+        ores.put(new OreType(Material.STONE, 5, 0, 80, 32), 10);
+        ores.put(new OreType(Material.COAL_ORE, 0, 0, 128, 16), 20);
+        ores.put(new OreType(Material.IRON_ORE, 0, 0, 64, 8), 20);
+        ores.put(new OreType(Material.GOLD_ORE, 0, 0, 32, 8), 2);
+        ores.put(new OreType(Material.REDSTONE_ORE, 0, 0, 16, 7), 8);
+        ores.put(new OreType(Material.DIAMOND_ORE, 0, 0, 16, 7), 1);
+        ores.put(new OreType(Material.LAPIS_ORE, 0, 16, 16, 6), 1);
     }
 
     @Override
     public void populate(World world, Random random, Chunk chunk) {
-        generator.generate(world, random, chunk);
+        final int cx = (chunk.getX() << 4);
+        final int cz = (chunk.getZ() << 4);
+
+        for (Entry<OreType, Integer> entry : ores.entrySet()) {
+
+            final OreType oreType = entry.getKey();
+            for (int n = 0; n < entry.getValue(); n++) {
+
+                int sourceX = cx + random.nextInt(16);
+                int sourceZ = cz + random.nextInt(16);
+                int sourceY = oreType.getMinY() == oreType.getMaxY() ?
+                        random.nextInt(oreType.getMinY()) + random.nextInt(oreType.getMinY()) :
+                            random.nextInt(oreType.getMaxY() - oreType.getMinY()) + oreType.getMinY();
+
+                new OreVein(oreType).generate(world, random, sourceX, sourceY, sourceZ);
+            }
+        }
     }
 }
